feat: add stripe for user subscription

+++ api/src/services/user.service.js
@@ -64,6 +64,9 @@ const updateUserById = async (userId, updateBody) => {
   if (updateBody.email && (await User.isEmailTaken(updateBody.email, userId))) {
     throw new ApiError(httpStatus.BAD_REQUEST, 'Email already taken');
   }
+  if (updateBody.name) {
+    updateBody.avatar = `https://ui-avatars.com/api/?name=${updateBody.name}&background=random`;
+  }
   Object.assign(user, updateBody);
   await user.save();
   return user;
@@ -95,7 +98,7 @@ const uploadAvatar = async (userId, file) => {
     throw new ApiError(httpStatus.NOT_FOUND, 'User not found');
   }
   const name = strings.generateFileName();
-  const fileBuffer = await sharp(file.buffer).resize({ height: 128, width: 128 }).toBuffer();
+  const fileBuffer = await sharp(file.buffer).resize({ height: 256, width: 256 }).toBuffer();
   await uploadService.uploadFile({
     ...file,
     buffer: fileBuffer,
@@ -117,11 +120,11 @@ const deleteAvatar = async (userId) => {
   if (!user) {
     throw new ApiError(httpStatus.NOT_FOUND, 'User not found');
   }
-  if (user.avatar !== null) {
+  if (user.avatar !== null && user.avatar.includes('amazonaws.com')) {
     const key = user.avatar.split('/').pop();
     await uploadService.deleteFile(key);
   }
-  user.avatar = null;
+  user.avatar = `https://ui-avatars.com/api/?name=${user.name}&background=random`;
   await user.save();
   return user;
 };
